pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "fixer-prototype"
description = "A prototype demonstrating how to implement configurable fixes"
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.12"
dependencies = ["fixer", "fixer-cmip7", "fixer-esa-cci"]
[project.optional-dependencies]
test = ["pytest"]
dev = ["pre-commit", "netcdf4", "dask", "scitools-iris"]

[tool.setuptools_scm]
version_scheme = "release-branch-semver"

[tool.uv]
required-version = ">=0.9.0"

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
fixer = { workspace = true }
fixer-cmip7 = { workspace = true }
fixer-esa-cci = { workspace = true }
